refactor: move CDeterministicMN::ToJson to core_write to avoid g_txindex dependency for linking
